+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - File viewer in chats (Image/Animated Image/Video/Texts) with the option to export the file.
 - Accept/decline CLI commands.
 - Colored Emojis with its own font family.
+- OAuth2 connection to retrieve remote provisioning (Experimental and not usable without configuration).
 
 ## 5.0.11 - undefined